// src/types/Sanity/contact.ts
export interface SanityContactData {
  cardHeading: string;
  cardSubheading: string;
  profileName: string;
  profileEmail: string;
  profilePhone: string;
  profileImage: string;
  formHeading: string;
  nameLabel: string;
  namePlaceholder: string;
  companyLabel: string;
  companyPlaceholder: string;
  emailLabel: string;
  emailPlaceholder: string;
  phoneLabel: string;
  phonePlaceholder: string;
  companySizeLabel: string;
  companySizePlaceholder: string;
  commentsLabel: string;
  commentsPlaceholder: string;
  formSubmitLabel: string;
  consentText: string;
}